refactor: reorganize pricelist widget
Move pricelist widget file into the pricelist directory
Split content and style controls into separate modular files
Update widget manager to load new control directories
